+++ packages/nitro/src/runtime/server/index.ts
@@ -2,11 +2,11 @@ import '../app/config'
 import { createApp, useBase } from 'h3'
 import { createFetch } from 'ohmyfetch'
 import destr from 'destr'
-import { createCall, createFetch as createLocalFetch } from 'unenv/runtime/fetch'
+import { createCall, createFetch as createLocalFetch } from 'unenv/runtime/fetch/index'
 import { timingMiddleware } from './timing'
 import { handleError } from './error'
 // @ts-ignore
-import serverMiddleware from '~serverMiddleware'
+import serverMiddleware from '#server-middleware'
 
 const app = createApp({
   debug: destr(process.env.DEBUG),