+++ playground/pages/index.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    Hello
+    Hello World
   </div>
 </template>