+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.8](https://github.com/nuxt/framework/compare/nuxt-cli@0.3.7...nuxt-cli@0.3.8) (2021-06-24)
+
+**Note:** Version bump only for package nuxt-cli
+
+
+
+
+
+## [0.3.7](https://github.com/nuxt/framework/compare/nuxt-cli@0.3.6...nuxt-cli@0.3.7) (2021-06-16)
+
+**Note:** Version bump only for package nuxt-cli
+
+
+
+
+
 ## [0.3.6](https://github.com/nuxt/framework/compare/nuxt-cli@0.3.5...nuxt-cli@0.3.6) (2021-06-04)
 
 **Note:** Version bump only for package nuxt-cli